Move kickerPicker to its own file.
